+++ main/webui/app.js
@@ -972,7 +972,7 @@
                 mqttWarningsTopic: "paradox/tfd/dynamite/warnings",
                 mqttGameStateTopic: "paradox/tfd/state",
                 mqttPropAnnounceTopic: "/Paradox/Props",
-                networkName: "px-dynamite-v1-a1b2",
+                networkName: "dynamite",
                 apSsid: "Paradox-PXDynamiteV1-A1B2",
                 apIpAddress: "192.168.4.1",
                 apPassword: "",